ANDDEP-255 Fix

+++ template/app-injector/src/main/java/ru/surfstudio/standard/app_injector/network/OkHttpModule.kt
@@ -14,7 +14,7 @@ import java.util.concurrent.TimeUnit
 import javax.inject.Named
 
 const val DI_NAME_SERVICE_INTERCEPTOR = "DI_NAME_SERVICE_INTERCEPTOR"
-private const val NETWORK_TIMEOUT = 30 //sec
+private const val NETWORK_TIMEOUT = 30L //sec
 
 /**
  * этот модуль вынесен отдельно для возможности замены его при интеграционном тестировании
@@ -36,9 +36,9 @@ class OkHttpModule {
                                      etagInterceptor: EtagInterceptor,
                                      httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
         val okHttpClientBuilder = OkHttpClient.Builder()
-        okHttpClientBuilder.connectTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-        okHttpClientBuilder.readTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-        okHttpClientBuilder.writeTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
+        okHttpClientBuilder.connectTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+        okHttpClientBuilder.readTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+        okHttpClientBuilder.writeTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
 
         okHttpClientBuilder.addInterceptor(cacheInterceptor)
         okHttpClientBuilder.addInterceptor(etagInterceptor)
